+++ web/src/enterprise/site-admin/SiteAdminLsifJobPage.tsx
@@ -1,19 +1,19 @@
-import * as GQL from '../../../shared/src/graphql/schema'
+import * as GQL from '../../../../shared/src/graphql/schema'
 import AlertCircleIcon from 'mdi-react/AlertCircleIcon'
 import CheckIcon from 'mdi-react/CheckIcon'
 import ClockOutlineIcon from 'mdi-react/ClockOutlineIcon'
 import React, { FunctionComponent, useEffect, useMemo } from 'react'
-import { asError, ErrorLike, isErrorLike } from '../../../shared/src/util/errors'
+import { asError, ErrorLike, isErrorLike } from '../../../../shared/src/util/errors'
 import { catchError } from 'rxjs/operators'
-import { eventLogger } from '../tracking/eventLogger'
-import { fetchLsifJob } from './backend'
+import { eventLogger } from '../../tracking/eventLogger'
+import { fetchLsifJob } from '../../site-admin/backend'
 import { LoadingSpinner } from '@sourcegraph/react-loading-spinner'
-import { PageTitle } from '../components/PageTitle'
+import { PageTitle } from '../../components/PageTitle'
 import { sortBy } from 'lodash'
 import { RouteComponentProps } from 'react-router'
-import { Timestamp } from '../components/time/Timestamp'
-import { useObservable } from '../util/useObservable'
-import { ErrorAlert } from '../components/alerts'
+import { Timestamp } from '../../components/time/Timestamp'
+import { useObservable } from '../../util/useObservable'
+import { ErrorAlert } from '../../components/alerts'
 
 const JobArguments: FunctionComponent<{ args: { [name: string]: string } }> = ({ args }) => (
     <table className="job-arguments w-100 mb-0 table table-sm">
